moved tree state into context
