+++ algorithms/Trees/BST.ts
