+++ pages/_app.tsx
@@ -3,6 +3,7 @@ import { NextUIProvider, createTheme, Grid } from "@nextui-org/react";
 import { ThemeProvider } from "next-themes";
 import { Header, Layout, Sidebar } from "@components/Layout/index";
 import { useRouter } from "next/router";
+import BSTContextProvider from "@contexts/BST";
 
 function MyApp({ Component, pageProps }: AppProps) {
   const router = useRouter();
@@ -25,21 +26,23 @@ function MyApp({ Component, pageProps }: AppProps) {
       }}
     >
       <NextUIProvider>
-        <Layout>
-          <Header />
-          {showSidebar ? (
-            <Grid.Container gap={2}>
-              <Grid xs={12} sm={2.5}>
-                <Sidebar />
-              </Grid>
-              <Grid xs={12} sm={9.5}>
-                <Component {...pageProps} />
-              </Grid>
-            </Grid.Container>
-          ) : (
-            <Component {...pageProps} />
-          )}
-        </Layout>
+        <BSTContextProvider>
+          <Layout>
+            <Header />
+            {showSidebar ? (
+              <Grid.Container gap={2}>
+                <Grid xs={12} sm={2.5}>
+                  <Sidebar />
+                </Grid>
+                <Grid xs={12} sm={9.5}>
+                  <Component {...pageProps} />
+                </Grid>
+              </Grid.Container>
+            ) : (
+              <Component {...pageProps} />
+            )}
+          </Layout>
+        </BSTContextProvider>
       </NextUIProvider>
     </ThemeProvider>
   );
